PAGI-158 refactored tests for validation
